Add example folder + fix broken scoped slots in components + fix eslint config
